apps/backend/pyproject.toml
[tool.poetry]
name = "itx-backend"
version = "0.1.0"
description = "Backend for IncomeTax Agent"
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
langgraph = "*"
psycopg2-binary = "*"
pydantic = "^2.0.0"
opentelemetry-sdk = "*"
opentelemetry-api = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
